+++ core/database/src/commonMain/kotlin/com/erp/jytextile/core/database/model/FabricRollEntity.kt
@@ -24,8 +24,8 @@ data class FabricRollEntity(
     val code: String,
     val color: String,
     val factory: String,
-    @ColumnInfo("remaining_length") val remainingLength: Int,
-    @ColumnInfo("original_length") val originalLength: Int,
+    @ColumnInfo("remaining_length") val remainingLength: Double,
+    @ColumnInfo("original_length") val originalLength: Double,
     val remark: String?,
 )
